# backend/docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ALPACART API",
    "description": "API Backend de Alpacart — Marca premium de alpaca peruana",
    "version": "1.0.0",
    "contact": {
      "name": "Alpacart Textiles",
      "url": "https://alpacart.com"
    }
  },
  "servers": [
    { "url": "http://localhost:8000/api/v1", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/cart": {
      "get": {
        "tags": ["Customers"],
        "summary": "Mi carrito",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Carrito con items" } }
      },
      "delete": {
        "tags": ["Customers"],
        "summary": "Vaciar carrito",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Carrito vaciado" } }
      }
    },
    "/cart/items": {
      "post": {
        "tags": ["Customers"],
        "summary": "Agregar al carrito",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": { "type": "string", "format": "uuid" },
                  "variantId": { "type": "string", "format": "uuid" },
                  "quantity": { "type": "integer", "minimum": 1 }
                },
                "required": ["productId", "quantity"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Item agregado" } }
      }
    },
    "/cart/items/{id}": {
      "patch": {
        "tags": ["Customers"],
        "summary": "Actualizar cantidad",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Cantidad actualizada" } }
      },
      "delete": {
        "tags": ["Customers"],
        "summary": "Eliminar item del carrito",
        "security": [{ "bearer": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Item eliminado" } }
      }
    },
    "/checkout": {
      "post": {
        "tags": ["Customers"],
        "summary": "Procesar checkout",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "couponCode": { "type": "string" },
                  "idempotencyKey": { "type": "string", "description": "Idempotency-Key header" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Orden creada" } }
      }
    },
...
    "/wishlist": {
      "get": {
        "tags": ["Customers"],
        "summary": "Lista de deseos",
        "security": [{ "bearer": [] }],
        "responses": { "200": { "description": "Lista de deseos" } }
      }
    },
    "/wishlist/items": {
      "post": {
        "tags": ["Customers"],
        "summary": "Agregar/quitar de wishlist",
        "security": [{ "bearer": [] }],
        "responses": { "201": { "description": "Item agregado/removido" } }
      }
    },
...
  }
}